Coletar dados de orienta7 de voto no Senado Federal

Esta fun7 coleta os dados de orienta7 das vota7 do Senado Federal a partir de uma lista de datas de sess.

Usage

coletar_orientacao_votacao(data_sessao)

Arguments

data_sessao

Um vetor contendo as datas das sess no formato "AAAA-MM-DD".

Value

Um data frame contendo os dados de vota7, incluindo c de vota7, data e hora, partido e voto.

Examples

anos <- c(2022, 2023)
dados_votacoes <- extrair_votacoes_por_ano(anos)
dados_votacao <- coletar_orientacao_votacao(head(dados_votacoes$data_sessao))