Extrair detalhes de vetos a partir da URL específica contendo informações sobre vetos

Esta função realiza scraping de uma página contendo informações sobre vetos e extrai detalhes como número do veto, link associado, ementa, data de sobresta, matéria vetada e norma gerada.

extrair_detalhes_vetos(pages = 1)

Arguments

pages

O número de páginas a serem processadas.

Value

Um dataframe contendo os detalhes dos vetos.

Examples

dados_vetos <- extrair_detalhes_vetos(pages = 3)
head(dados_vetos)
#>                    Veto
#> 1   VET 38/2023 - Total
#> 2   VET 34/2023 - Total
#> 3 VET 33/2023 - Parcial
#> 4 VET 32/2023 - Parcial
#> 5 VET 31/2023 - Parcial
#> 6 VET 30/2023 - Parcial
#>                                                                       Link
#> 1 https://www.congressonacional.leg.br/materias/vetos/-/veto/detalhe/16110
#> 2 https://www.congressonacional.leg.br/materias/vetos/-/veto/detalhe/16049
#> 3 https://www.congressonacional.leg.br/materias/vetos/-/veto/detalhe/16029
#> 4 https://www.congressonacional.leg.br/materias/vetos/-/veto/detalhe/16009
#> 5 https://www.congressonacional.leg.br/materias/vetos/-/veto/detalhe/15989
#> 6 https://www.congressonacional.leg.br/materias/vetos/-/veto/detalhe/15969
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                Ementa
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             Veto Total aposto ao Projeto de Lei nº 334, de 2023, que "Prorroga até 31 de dezembro de 2027 os prazos de que tratam os arts. 7º e 8º da Lei nº 12.546, de 14 de dezembro de 2011, e o "caput" do  21 do art. 8º da Lei nº 10.865, de 30 de abril de 2004, e dá outras providências".
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           Veto Total aposto ao Projeto de Lei nº 1.949, de 2021, que "Acresce dispositivo à Consolidação das Leis do Trabalho, aprovada pelo Decreto-Lei nº 5.452, de 1º de maio de 1943, para estabelecer a não caracterização como perigosas das atividades ou operações que envolvam exposição às quantidades de inflamáveis contidas em tanques de combustíveis originais de fábrica e suplementares, na forma que especifica".
#> 3 Veto Parcial aposto ao Projeto de Lei nº 4.188, de 2021, que "Dispõe sobre o aprimoramento das regras de garantia, a execução extrajudicial de créditos garantidos por hipoteca, a execução extrajudicial de garantia imobiliária em concurso de credores, o procedimento de busca e apreensão extrajudicial de bens móveis em caso de inadimplemento de contrato de alienação fiduciária, o resgate antecipado de Letra Financeira, a alíquota de imposto de renda sobre rendimentos no caso de fundos de investimento em participações qualificados que envolvam titulares de cotas com residência ou domicílio no exterior e o procedimento de emissão de debêntures; altera as Leis nºs 9.514, de 20 de novembro de 1997, 10.406, de 10 de janeiro de 2002 (Código Civil), 13.476, de 28 de agosto de 2017, 6.015, de 31 de dezembro de 1973 (Lei de Registros Públicos), 6.766, de 19 de dezembro de 1979, 13.105, de 16 de março de 2015 (Código de Processo Civil), 9.492, de 10 de setembro de 1997, 8.935, de 18 de novembro de 1994, 12.249, de 11 de junho de 2010, 14.113, de 25 de dezembro de 2020, 11.312, de 27 de junho de 2006, 6.404, de 15 de dezembro de 1976, e 14.382, de 27 de junho de 2022, e o Decreto-Lei nº 911, de 1º de outubro de 1969; e revoga dispositivos dos Decretos-Lei nºs 70, de 21 de novembro de 1966, e 73, de 21 de novembro de 1966".
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     Veto Parcial aposto ao Projeto de Lei nº 5.614 de 2020 (nº 9.382/2017, na Câmara dos Deputados), que "Altera a Lei nº 12.319, de 1º de setembro de 2010, para dispor sobre o exercício profissional e as condições de trabalho do profissional tradutor, intérprete e guia-intérprete da Língua Brasileira de Sinais (Libras)".
#> 5                                                                        Veto Parcial aposto ao Projeto de Lei Complementar nº 136, de 2023, que "Dispõe sobre a compensação devida pela União nos termos dos arts. 3º e 14 da Lei Complementar nº 194, de 23 de junho de 2022, a dedução das parcelas dos contratos de dívida, a transferência direta de recursos da União aos Estados e ao Distrito Federal, a incorporação do excesso compensado judicialmente em saldo devedor de contratos de dívida administrados pela Secretaria do Tesouro Nacional do Ministério da Fazenda, o tratamento jurídico e contábil aplicável aos pagamentos, às compensações e às vinculações, as transferências de recursos aos Municípios em razão da redução das receitas do Fundo de Participação dos Municípios (FPM), as transferências de recursos aos Estados e ao Distrito Federal em razão da redução das receitas do Fundo de Participação dos Estados e do Distrito Federal (FPE) e as regras relativas ao Imposto sobre Operações Relativas à Circulação de Mercadorias e sobre Prestações de Serviços de Transporte Interestadual e Intermunicipal e de Comunicação (ICMS); e revoga dispositivos da Lei nº 5.172, de 25 de outubro de 1966 (Código Tributário Nacional), e das Leis Complementares nºs 87, de 13 de setembro de 1996 (Lei Kandir), e 192, de 11 de março de 2022".
#> 6                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            Veto Parcial aposto ao Projeto de Lei nº 2.903, de 2023 (nº 490/2007, na Câmara dos Deputados), que "Regulamenta o art. 231 da Constituição Federal, para dispor sobre o reconhecimento, a demarcação, o uso e a gestão de terras indígenas; e altera as Leis nºs 11.460, de 21 de março de 2007, 4.132, de 10 de setembro de 1962, e 6.001, de 19 de dezembro de 1973".
#>     Sobresta Materia_Vetada                                   Norma_Gerada
#> 1 03/02/2024    PL 334/2023 Lei nº 14.784 de 27/12/2023     Veto rejeitado
#> 2 06/12/2023   PL 1949/2021 Lei nº 14.766 de 22/12/2023     Veto rejeitado
#> 3 30/11/2023   PL 4188/2021                    Lei nº 14.711 de 30/10/2023
#> 4 25/11/2023   PL 5614/2020                    Lei nº 14.704 de 25/10/2023
#> 5 24/11/2023   PLP 136/2023          Lei Complementar nº 201 de 24/10/2023
#> 6 22/11/2023   PL 2903/2023                    Lei nº 14.701 de 20/10/2023