Extrai Discursos Parlamentares de um ou mais Senadores

Esta fun7 extrai os discursos parlamentares de um ou mais senadores em um intervalo de datas fornecido a partir da API do Senado.

Usage

extrair_discursos(codigo_senador, data_inicio, data_fim)

Arguments

codigo_senador

Um vetor de c de senadores. Exemplo: c("5529", "1234").

data_inicio

A data de in do intervalo, no formato "YYYY-MM-DD". Exemplo: "2024-01-01".

data_fim

A data de fim do intervalo, no formato "YYYY-MM-DD". Exemplo: "2024-06-30".

Value

Um dataframe com os discursos parlamentares, contendo informa7 como c do discurso, data, partido, UF, e resumo.

Examples

# Exemplo de uso
discursos_df <- extrair_discursos(codigo_senador = c("5529"), data_inicio = "2024-01-01", data_fim = "2024-06-30")
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union