Extrair dados de reuni de comiss do Senado

Esta fun7 extrai dados de reuni de comiss do Senado Federal com base nos c das comiss e no intervalo de datas fornecido.

Usage

info_dados_reuniao_comissao(codcol, data_inicio, data_fim)

Arguments

codcol

Vetor contendo os c das comiss

data_inicio

Data de in no formato "YYYY-MM-DD"

data_fim

Data de fim no formato "YYYY-MM-DD"

Value

Um DataFrame com os dados das reuni das comiss

Examples

codcol <- c(54, 38, 2614, 34)
data_inicio <- "2023-01-01"
data_fim <- "2023-12-01"
df_reunioes <- info_dados_reuniao_comissao(codcol, data_inicio, data_fim)