Extrair detalhes de vetos a partir da URL espec contendo informa7 sobre vetos

Esta fun7 realiza scraping de uma p contendo informa7 sobre vetos e extrai detalhes como n do veto, link associado, ementa, data de sobresta, mat vetada e norma gerada.

Usage

info_vetos(pages = 1)

Arguments

pages

O n de p a serem processadas. Padr 9 1.

Value

Um dataframe contendo os detalhes dos vetos.

Examples

# Exemplo de uso
dados_vetos <- info_vetos(pages = 3)