Processa XML de Apartes de Senadores

Esta fun7 acessa a URL da API do Senado para cada c de parlamentar fornecido, extrai e processa os dados dos apartes associados a cada parlamentar. Retorna um dataframe com informa7 sobre os apartes, incluindo dados do parlamentar e detalhes dos apartes.

Usage

processar_xml_apartes(codigos)

Arguments

codigos

Um vetor de c de parlamentares para os quais os dados de apartes ser extra.

Value

Um dataframe contendo os dados dos apartes, com as seguintes colunas:

CodigoParlamentar

C do parlamentar

NomeParlamentar

Nome do parlamentar

CodigoPronunciamento

C do pronunciamento

TipoUsoPalavra

Tipo de uso da palavra

DataPronunciamento

Data do pronunciamento

SiglaPartidoParlamentarNaData

Sigla do partido do parlamentar na data

UfParlamentarNaData

UF do parlamentar na data

SiglaCasaPronunciamento

Sigla da casa do pronunciamento

NomeCasaPronunciamento

Nome da casa do pronunciamento

TextoResumo

Resumo do texto do pronunciamento

Indexacao

Indexa7 do pronunciamento

UrlTexto

URL do texto do pronunciamento

UrlTextoBinario

URL do texto bin do pronunciamento

CodigoSessao

C da sess plen

NomeCasaSessao

Nome da casa da sess

DataSessao

Data da sess

HoraInicioSessao

Hora de in da sess

OradorCodigo

C do orador

OradorNome

Nome do orador

PublicacaoData

Data da publica7

PublicacaoUrl

URL da publica7

Examples

if (FALSE) { # \dontrun{
# Exemplo de uso com c\u00f3digos fict\u00edcios
codigos <- c("4763", "1234")
df_apartes <- processar_xml_apartes(codigos)
} # }